Consolidate type aliases into a single file
