+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/config/behavior/DataCaptureEventBehavior.kt
@@ -2,6 +2,7 @@ package io.embrace.android.embracesdk.config.behavior
 
 import io.embrace.android.embracesdk.config.remote.RemoteConfig
 import io.embrace.android.embracesdk.internal.PatternCache
+import io.embrace.android.embracesdk.internal.utils.UnimplementedConfig
 
 internal class DataCaptureEventBehavior(
     thresholdCheck: BehaviorThresholdCheck,
