+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/config/behavior/LogMessageBehavior.kt
@@ -1,6 +1,7 @@
 package io.embrace.android.embracesdk.config.behavior
 
 import io.embrace.android.embracesdk.config.remote.LogRemoteConfig
+import io.embrace.android.embracesdk.internal.utils.UnimplementedConfig
 
 /**
  * Provides the behavior that should be followed for remote log message functionality.
