--- embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/config/behavior/UnimplementedConfig.kt
@@ -1,6 +1,0 @@
-package io.embrace.android.embracesdk.config.behavior
-
-/**
- * When a local or remote config doesn't exist, this type can be used as a placeholder.
- */
-internal typealias UnimplementedConfig = Unit?
