+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/FakeApiClient.kt
@@ -3,7 +3,7 @@ package io.embrace.android.embracesdk.fakes
 import io.embrace.android.embracesdk.comms.api.ApiClient
 import io.embrace.android.embracesdk.comms.api.ApiRequest
 import io.embrace.android.embracesdk.comms.api.ApiResponse
-import io.embrace.android.embracesdk.comms.api.SerializationAction
+import io.embrace.android.embracesdk.internal.utils.SerializationAction
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.util.LinkedList
