+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/FakePendingApiCallsSender.kt
@@ -2,8 +2,8 @@ package io.embrace.android.embracesdk.fakes
 
 import io.embrace.android.embracesdk.comms.api.ApiRequest
 import io.embrace.android.embracesdk.comms.api.ApiResponse
-import io.embrace.android.embracesdk.comms.api.SerializationAction
 import io.embrace.android.embracesdk.comms.delivery.PendingApiCallsSender
+import io.embrace.android.embracesdk.internal.utils.SerializationAction
 import java.io.ByteArrayOutputStream
 import java.util.LinkedList
 import java.util.Queue
